Rename Transformers and move ServingFetcher
