+++ r2-streamer/src/main/java/org/readium/r2/streamer/server/Server.kt
@@ -16,7 +16,6 @@ import org.readium.r2.shared.Injectable
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.streamer.BuildConfig.DEBUG
 import org.readium.r2.streamer.container.Container
-import org.readium.r2.streamer.fetcher.ServingFetcher
 import org.readium.r2.streamer.server.handler.*
 import timber.log.Timber
 import java.io.File
@@ -177,7 +176,11 @@ abstract class AbstractServer(private var port: Int) : RouterNanoHTTPD("127.0.0.
             return
         }
         val baseUrl = URL(Publication.localBaseUrlOf(filename = fileName, port = port))
-        val fetcher = ServingFetcher(publication, userPropertiesPath, customResources)
+        val fetcher = ServingFetcher(
+            publication,
+            userPropertiesPath,
+            customResources
+        )
 
         publication.setSelfLink("$baseUrl/manifest.json")
