+++ r2-streamer/src/main/java/org/readium/r2/streamer/server/handler/ManifestHandler.kt
@@ -17,7 +17,7 @@ import org.nanohttpd.protocols.http.response.Status
 import org.nanohttpd.router.RouterNanoHTTPD
 import org.readium.r2.shared.format.MediaType
 import org.readium.r2.streamer.BuildConfig.DEBUG
-import org.readium.r2.streamer.fetcher.ServingFetcher
+import org.readium.r2.streamer.server.ServingFetcher
 import timber.log.Timber
 import java.io.IOException
