+++ r2-streamer/src/test/resources/org/readium/r2/streamer/parser/epub/deobfuscation/cut-cut.obf.woff
